Add file and group to pbxproj.
