+++ Sources/TezosGenUtils/ContractCodeGenerator.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by Marek Fořt on 11/14/19.
+//
+
+import Foundation
